refactor(p2p): share collection token resolution
Keep embedded and HTTP P2P admin adapters on one name-or-ID resolution implementation so reconcile behavior cannot drift between clients.
